Enabled the icon request class to create some extra XMLs and performed the following changes:
* Increased the buffer size. I ran into issues with lots of apps, increasing the buffer helped
* Created a "drawables" and "icon_pack" set of xmls, which related to the xmls in the app. This helps devs implement the requests easier
* Delete the /designriftsrequest folder when we get to the handler, with a check on the thread being finished as a precaution
* Dumped the icons into an icons folder in the zip. Makes it way cleaner
* Added "Component{" to the app_filter stuff, so that doesn't need to be done by a dev
* Created a variable called imageName that generates a lowercase app name for the image, drawable and array xmls

That's all. It's still very quick to grab all the icons, and it makes it faster/easier to implement. The only issue I could foresee is the imageName variable being bad if there are two apps of the same name. But that would realisitically only happen to system apps, which rarely have the same name.